airflow_project/requirements-airflow.txt
# Airflow Trading Workflow Requirements
# Install with: pip install -r requirements-airflow.txt

# Core Airflow (compatible with Python 3.13)
apache-airflow==3.0.4
apache-airflow-providers-postgres==5.10.0

# Database dependencies
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0

# Data processing
pandas>=2.2.0
numpy>=2.0.0

# Web scraping dependencies
selenium>=4.15.0
webdriver-manager>=4.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Trading/Finance dependencies
ib-insync>=0.9.86

# Utility dependencies
python-dotenv>=0.19.0
coolname>=2.2.0

# Development/Testing
pytest>=7.0.0
black>=23.0.0

# Optional monitoring
prometheus-client>=0.19.0